fixed compile
impr. illegal reads
fix s16 reads
asm thumb emulation

+++ hyperspeedup/code_container/arm9/source/inlinetonotinline2.arm.cpp
@@ -45,11 +45,11 @@ void CPUWriteBytepuextern(u32 address, u8 b)
 	CPUWriteBytepu(address,b);
 }
 
-u16 CPUReadHalfWordrealpuSigned(u32 address);
+s32 CPUReadHalfWordrealpuSigned(u32 address);
 
-s16 CPUReadHalfWordrealpuSignedoutline(u32 address)
+s32 CPUReadHalfWordrealpuSignedoutline(u32 address)
  {
-	return (s16)CPUReadHalfWordrealpuSigned(address);
+	return CPUReadHalfWordrealpuSigned(address);
  }
 
 s8 CPUReadByteSignedpu(u32 address)
